Headless Enhancement
After more testing, found that the automation still wasn't working truly headlessly (as in screenlocked). These updates, while not the original planned approach, facilitate fully headless operation. But they do require an explicit parameter for most file/edit commands to specific the target document, to be fully safe. This parameter is optional, so a user not needing headless operation can skip this.